requirements.txt
click==8.1.3
easydict==1.10
ghapi==1.0.6
icecream==2.1.4
matplotlib==3.5.2
mayavi==4.8.1
nox==2025.2.9
numpy==1.18.1
open3d==0.17.0
Pillow==9.5.0
Pillow==6.1.0
Pillow==11.1.0
pytest==8.3.5
pytorch_lightning==1.5.10
PyYAML==6.0
PyYAML==6.0.2
rich==13.9.4
setuptools==59.5.0
spconv_cu113==2.3.6
timm==0.9.7
torch==1.10.1
torch_geometric==1.7.2
torch_scatter==2.0.9
tqdm==4.65.0
